+++ applications/dashboard/src/scripts/@types/api/general.ts
