+++ applications/dashboard/src/scripts/@types/api/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license https://opensource.org/licenses/GPL-2.0 GPL-2.0
+ */
+
+export * from "@dashboard/@types/api/general";
+export * from "@dashboard/@types/api/user";
+export * from "@dashboard/@types/api/authenticate";
