+++ applications/dashboard/src/scripts/@types/api/user.ts
