+++ applications/dashboard/src/scripts/@types/state.ts
@@ -5,8 +5,8 @@
  * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
  */
 
-import { ILoadable } from "@dashboard/types/api";
-import { IUserAuthenticator } from "@dashboard/types/api/authenticate";
+import { ILoadable } from "@dashboard/@types/api";
+import { IUserAuthenticator } from "@dashboard/@types/api/authenticate";
 
 export type IAuthenticatorState = ILoadable<IUserAuthenticator[]>;
