+++ applications/dashboard/src/scripts/@types/store/IAuthenticateState.ts
