+++ applications/dashboard/src/scripts/state/session/authenticatorsActions.ts
@@ -7,8 +7,8 @@ import { Dispatch } from "redux";
 import { generateApiActionCreators, ActionsUnion, createAction } from "@dashboard/state/utility";
 import api from "@dashboard/apiv2";
 import { AxiosResponse } from "axios";
-import { LoadStatus, IUserAuthenticator } from "@dashboard/types/api";
-import { IStoreState } from "@dashboard/types/state";
+import { LoadStatus, IUserAuthenticator } from "@dashboard/@types/api";
+import { IStoreState } from "@dashboard/@types/state";
 
 export const GET_SIGNIN_AUTHENTICATORS_REQUEST = "GET_SIGNIN_AUTHENTICATORS_REQUEST";
 export const GET_SIGNIN_AUTHENTICATORS_ERROR = "GET_SIGNIN_AUTHENTICATORS_ERROR";
@@ -22,7 +22,7 @@ const getAuthenticatorsActions = generateApiActionCreators(
     [] as IUserAuthenticator[],
 );
 
-export function getAuthenticators() {
+export function getUserAuthenticators() {
     return (dispatch: Dispatch<any>, getState: () => IStoreState) => {
         dispatch(getAuthenticatorsActions.request());
